masking: image difference

+++ analysis/image_capture.py
@@ -102,7 +102,7 @@ def applySaliencyMap(img_array):
 
     image_out = []
     for img in img_array:
-        gray = cv2.cvtColor(np.array(img), cv2.COLOR_BGR2GRAY)
+        gray = cv2.cvtColor(np.array(img), cv2.COLOR_BGR2GRAY).astype(np.float32)
         print("0:", gray)
         (success, saliencyMap) = saliency.computeSaliency(gray)
         print("success:", success)
